Support the config of HTTP proxy in anonymous & basic auth
Issue #3.

Actually, there are many schemes of authentication, including basic,
digest, NTLM, and so on.

In this moment, I have only added support of the basic auth.
You can modify the following line to fit your need and add it to your
config file:

    HTTP_PROXY = "http://user:pass@localhost:3128/"

Please me know if it doesn't work.

By the way, define an environment variable is another way to use proxy,
to use this method even no need for this patch.

Export the variable before you running cf.py:

    export http_proxy=http://user:pass@localhost:3128/

or define the variable when you running:

    http_proxy=http://user:pass@localhost:3128/ cf.py ....